bug: implement function returning rides total cost
Before:
The rides page was showing an incorrect value for the rides total cost.
This value was taking into account only the rides being shown on
the current page. The cost of rides on the next pages was ignored.

After:
The rides page shows the total cost for all rides matching the filter
options, including those rides which are on the next pages.