feat: implement validation attribute class
Attribute Required was removed from the property Date of the class Ride.
A validation attribute called PastDate was implemented. It checks
whether the date has been filled out and it is not a past date.

Refactoring: initialMileage validation commands were moved to a method
called ValidateMileage on the RideController class.